Fix : added field member for tower's name and level
fixed name of script
CamoTurretLV1 -> CanonTurretLV1
CamoTurretLV2 -> CanonTurretLV2
CamoTurretLV3 -> CanonTurretLV3
